pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "This web app can help you to watch over your tasks."
authors = ["Iaroslav <GunGalla@gmail.com>"]
readme = "README.md"
packages = [
    {include = "task_manager"},
]
repository = "https://github.com/GunGalla/python-project-52"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
python-dotenv = "^1.0.0"
django-extensions = "^3.2.1"
dj-database-url = "0.5.0"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"
ipython = "^8.11.0"
django-filter = "^22.1"
rollbar = "^0.16.3"
psycopg2-binary = "^2.9.5"
coverage = "^7.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
